refactor: 구현 요구사항을 맞추기 위해 this.play() 함수의 return값을 String으로 바꿈
기존 gameProgress에서 결과까지 화면에 출력해주던 것을 분리하기 위해 기존 함수들도 정리
