+++ src/index.js
@@ -1,22 +1,31 @@
-import gameProgress from './game/gameProgress.js';
 import getUserInput from './input/getUserInput.js';
 import getRandomNumbers from './utils/getRandomNumbers.js';
+import getGameResult from './game/getGameResult.js';
+
+/*
+ ** export default function BaseballGame() {
+ **  this.play = function (computerInputNumbers, userInputNumbers) {
+ **     return "결과 값 String";
+ **  };
+ ** }
+ */
 
 export default function BaseballGame() {
-  // 이벤트 리스너 등록
   const $submit = document.querySelector('#submit');
+  const $app = document.querySelector('#app');
 
   const computerInputNumbers = getRandomNumbers();
 
   this.play = function (computerInputNumbers, userInputNumbers) {
-    return gameProgress(computerInputNumbers, userInputNumbers);
+    return getGameResult(computerInputNumbers, userInputNumbers);
   };
 
   $submit.addEventListener('click', () => {
     const userInputNumbers = getUserInput();
 
     if (userInputNumbers) {
-      this.play(computerInputNumbers, userInputNumbers);
+      const gameResult = this.play(computerInputNumbers, userInputNumbers);
+      console.log(typeof gameResult, gameResult);
     }
   });
 }
